keeping alive when routing
