+++ lib/config/router/app_router.dart
@@ -1,11 +1,14 @@
 import 'package:go_router/go_router.dart';
 import 'package:cuevanax/presentations/screens/screens.dart';
 
-final appRouter = GoRouter(initialLocation: '/', routes: [
+final appRouter = GoRouter(initialLocation: '/home/0', routes: [
   GoRoute(
-      path: '/',
+      path: '/home/:page',
       name: HomeScreen.name,
-      builder: (context, state) => const HomeScreen(),
+      builder: (context, state) {
+        final pageIndex = int.parse(state.pathParameters['page'] ?? '0');
+        return  HomeScreen(pageIndex: pageIndex,);
+      },
       routes: [
         GoRoute(
             path: 'movie/:id',
