+++ lib/presentations/screens/movies/home_screen.dart
@@ -1,109 +1,23 @@
-import 'package:cuevanax/presentations/providers/providers.dart';
-import 'package:cuevanax/presentations/widgets/widgets.dart';
+import 'package:cuevanax/presentations/views/views.dart';
 import 'package:flutter/material.dart';
 
-import 'package:flutter_riverpod/flutter_riverpod.dart';
+import '../../widgets/widgets.dart';
 
 class HomeScreen extends StatelessWidget {
   static const name = 'home-screen';
+  final int pageIndex;
+  const HomeScreen({super.key, required this.pageIndex});
 
-  const HomeScreen({super.key});
-
-  @override
-  Widget build(BuildContext context) {
-    return const Scaffold(
-        body: _HomeView(), bottomNavigationBar: CustomBottomNavigation());
-  }
-}
-
-class _HomeView extends ConsumerStatefulWidget {
-  const _HomeView();
-
-  @override
-  _HomeViewState createState() => _HomeViewState();
-}
-
-class _HomeViewState extends ConsumerState<_HomeView> {
-  @override
-  void initState() {
-    super.initState();
-    ref.read(nowPlayingMoviesProvider.notifier).loadNextPage();
-    ref.read(popularMoviesProvider.notifier).loadNextPage();
-    ref.read(upcomingMoviesProvider.notifier).loadNextPage();
-    ref.read(topRatedMoviesProvider.notifier).loadNextPage();
-  }
+  final viewRoutes = const <Widget>[
+    HomeView(),
+    SizedBox(),
+    FavoritesView(),
+  ];
 
   @override
   Widget build(BuildContext context) {
-    final initalLoading = ref.watch(initialLoadingProvider);
-    if (initalLoading) return const FullScreenLoader();
-
-    final slideShowMovies = ref.watch(moviesSlideshowProvider);
-    final nowPlayingMovies = ref.watch(nowPlayingMoviesProvider);
-    final popularMovies = ref.watch(popularMoviesProvider);
-    final upcomingMovies = ref.watch(upcomingMoviesProvider);
-    final topRatedMovies = ref.watch(topRatedMoviesProvider);
-
-    if (slideShowMovies.isEmpty) {
-      return const Center(child: CircularProgressIndicator());
-    }
-
-    return Visibility(
-      visible: !initalLoading,
-      child: CustomScrollView(slivers: [
-        const SliverAppBar(
-          floating: true,
-          flexibleSpace: FlexibleSpaceBar(
-            title: CustomAppBar(),
-          ),
-        ),
-        SliverList(
-            delegate: SliverChildBuilderDelegate(
-          (context, index) {
-            return Column(
-              children: [
-                MoviesSlideshow(movies: slideShowMovies),
-                MovieHorizontalListView(
-                  movies: nowPlayingMovies,
-                  title: 'In Cinema',
-                  subtitle: 'Monday ',
-                  loadNextPage: () {
-                    ref.read(nowPlayingMoviesProvider.notifier).loadNextPage();
-                  },
-                ),
-                MovieHorizontalListView(
-                  movies: popularMovies,
-                  title: 'Populars',
-                  subtitle: 'Best movies ',
-                  loadNextPage: () {
-                    ref.read(popularMoviesProvider.notifier).loadNextPage();
-                  },
-                ),
-                MovieHorizontalListView(
-                  movies: upcomingMovies,
-                  title: 'Upcoming',
-                  subtitle: 'In this Month ',
-                  loadNextPage: () {
-                    ref.read(upcomingMoviesProvider.notifier).loadNextPage();
-                  },
-                ),
-                MovieHorizontalListView(
-                  movies: topRatedMovies,
-                  title: 'Top Rated',
-                  subtitle: 'Most Ranked ',
-                  loadNextPage: () {
-                    ref.read(topRatedMoviesProvider.notifier).loadNextPage();
-                  },
-                ),
-                const SizedBox(
-                  height: 10,
-                )
-              ],
-            );
-          },
-          childCount: 1,
-        ))
-      ]),
-    );
+    return  Scaffold(
+        body: IndexedStack(index: pageIndex, children: viewRoutes), 
+        bottomNavigationBar: const CustomBottomNavigation());
   }
 }
