+++ lib/presentations/views/views.dart
@@ -1,0 +1,2 @@
+export 'package:cuevanax/presentations/views/movies/home_view.dart';
+export 'package:cuevanax/presentations/views/movies/favorites_view.dart';
